Split org.ccjmne.orca.api.rest into 5 more specific subpackages
