+++ src/main/java/org/ccjmne/orca/api/rest/edit/EmployeesNotesEndpoint.java
@@ -1,4 +1,4 @@
-package org.ccjmne.orca.api.rest;
+package org.ccjmne.orca.api.rest.edit;
 
 import static org.ccjmne.orca.jooq.classes.Tables.EMPLOYEES;
 import static org.ccjmne.orca.jooq.classes.Tables.EMPLOYEES_VOIDINGS;
