+++ src/main/java/org/ccjmne/orca/api/rest/pub/DemoEndpoint.java
@@ -1,4 +1,4 @@
-package org.ccjmne.orca.api.rest;
+package org.ccjmne.orca.api.rest.pub;
 
 import java.text.DateFormat;
 import java.text.SimpleDateFormat;
